refactor: use tool input error for ingestion contracts
