+++ ai_core/tools/__init__.py
@@ -1,0 +1,5 @@
+"""Tooling helpers and error types for AI Core."""
+
+from .errors import InputError, ToolError
+
+__all__ = ["ToolError", "InputError"]
